Change directory structure of library source tree
