+++ include/sha3/internals/utils.hpp
@@ -1,14 +1,9 @@
 #pragma once
 #include <cassert>
-#include <charconv>
 #include <cstddef>
 #include <cstdint>
 #include <cstring>
-#include <iomanip>
-#include <random>
 #include <span>
-#include <sstream>
-#include <type_traits>
 
 // Utility ( or commonly used ) functions for SHA3 implementation
 namespace sha3_utils {
@@ -103,64 +98,6 @@ u64_words_to_le_bytes(std::span<const uint64_t, rate / 64> words,
     u64_to_le_bytes(words[off], bytes.subspan(off * 8, 8));
     off++;
   }
-}
-
-// Generates N -many random values of type T | N >= 0
-template<typename T>
-static inline void
-random_data(std::span<T> data)
-  requires(std::is_unsigned_v<T>)
-{
-  std::random_device rd;
-  std::mt19937_64 gen(rd());
-  std::uniform_int_distribution<T> dis;
-
-  const size_t len = data.size();
-  for (size_t i = 0; i < len; i++) {
-    data[i] = dis(gen);
-  }
-}
-
-// Given a bytearray of length N, this function converts it to human readable
-// hex string of length N << 1 | N >= 0
-inline const std::string
-to_hex(std::span<const uint8_t> bytes)
-{
-  std::stringstream ss;
-  ss << std::hex;
-
-  for (size_t i = 0; i < bytes.size(); i++) {
-    ss << std::setw(2) << std::setfill('0') << static_cast<uint32_t>(bytes[i]);
-  }
-
-  return ss.str();
-}
-
-// Given a hex encoded string of length 2*L, this routine can be used for
-// parsing it as a byte array of length L.
-//
-// Taken from
-// https://github.com/itzmeanjan/ascon/blob/603ba1f223ddd3a46cb0b3d31d014312d96792b5/include/utils.hpp#L120-L145
-inline std::vector<uint8_t>
-from_hex(std::string_view hex)
-{
-  const size_t hlen = hex.length();
-  assert(hlen % 2 == 0);
-
-  const size_t blen = hlen / 2;
-  std::vector<uint8_t> res(blen, 0);
-
-  for (size_t i = 0; i < blen; i++) {
-    const size_t off = i * 2;
-
-    uint8_t byte = 0;
-    auto sstr = hex.substr(off, 2);
-    std::from_chars(sstr.data(), sstr.data() + 2, byte, 16);
-
-    res[i] = byte;
-  }
-
-  return res;
 }
 
 }
